backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.2
psycopg2-binary==2.9.9
mysqlclient==2.2.4
pymongo==4.6.3
clickhouse-driver==0.2.7
redis==5.0.3
celery==5.3.6
kombu==5.3.4
Pillow==10.3.0
python-dotenv==1.0.1
drf-spectacular==0.27.2
factory-boy==3.3.0
pytest==8.1.1
pytest-django==4.8.0
pytest-cov==5.0.0
coverage==7.4.4
